Move parental consent API to sub-router
